collection, add, update, gson V4.2
