+++ main/java/ua/goit/moneybot/model/UserInformation.java
@@ -1,4 +1,4 @@
-package ua.goit.moneybot.user;
+package ua.goit.moneybot.model;
 
 public class UserInformation {
     String name = "User";
